auth-server/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
pydantic_settings
alembic
psycopg2-binary